+++ src/Raven.CodeAnalysis/Macros/ITokenTreeExpressionMacro.cs
@@ -1,0 +1,8 @@
+namespace Raven.CodeAnalysis.Macros;
+
+public interface ITokenTreeExpressionMacro : IMacroDefinition
+{
+    MacroKind IMacroDefinition.Kind => MacroKind.FreestandingExpression;
+
+    FreestandingMacroExpansionResult Expand(TokenTreeMacroContext context);
+}